+++ app/src/main/java/com/example/logintask1/domain/use_cases/GetPostsUseCase.kt
@@ -1,0 +1,10 @@
+package com.example.logintask1.domain.use_cases
+
+import com.example.logintask1.domain.repository.PostsRepositoryImpl
+import javax.inject.Inject
+
+class GetPostsUseCase @Inject constructor(repository: PostsRepositoryImpl) {
+    fun invoke(){
+
+    }
+}
