+++ app/src/main/java/com/example/logintask1/ui/home/capture/HomeViewModel.kt
@@ -1,10 +1,8 @@
 package com.example.logintask1.ui.home.capture
 
-import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.logintask1.domain.user.ListItem
 
 class HomeViewModel : ViewModel() {
     private val _personalPosts = MutableLiveData<List<ListItem>>()
