+++ app/src/main/java/com/example/logintask1/ui/home/userpost/PostsUiModel.kt
@@ -2,7 +2,6 @@ package com.example.logintask1.ui.home.userpost
 
 import android.net.Uri
 import com.example.logintask1.R
-import com.example.logintask1.domain.user.PostButtonStateIcons
 
 data class PostsUiModel(
     val postId: Int,
