+++ app/src/main/java/com/example/logintask1/ui/home/userpost/adapter/PostListAdapter.kt
@@ -5,8 +5,8 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
-import com.example.logintask1.domain.user.UserPost
-import com.example.logintask1.domain.user.toUiModel
+import com.example.logintask1.ui.home.userpost.UserPost
+import com.example.logintask1.ui.home.userpost.toUiModel
 import com.example.logintask1.databinding.PostBinding
